requirements.txt
keras==2.2.4
opencv-python
pyqt5
pillow
tensorflow_gpu==1.15